+++ src/tests/dict_test.rs
@@ -157,7 +157,7 @@ fn test_dict_decoding() {
             fail_counter_size += 1;
         }
 
-        if frame_dec.bytes_read_from_source() != file_size as u64 {
+        if frame_dec.bytes_read_from_source() != file_size {
             println!(
                 "Framedecoder counted wrong amount of bytes: {}, should be: {}",
                 frame_dec.bytes_read_from_source(),
